Add a new hash style io wrapper for testing
This IO wrapper internally stores the io offsets of each added file and
returns the content of a file as a string if the #[](path) method is
used. It also provides a #to_h method to return a hash where each key is
a path and each value the content of that path.