Add workbook tests
This adds mostly integration style tests which assert that the proper
files are written by using a spy variant of the hash io wrapper.
The spy is required because we can't currently inject an io wrapper
instance, only a class, so we inject a class that always returns the
same instance.