Replace RubyZip with ZipTricks (#16)
This replaces the default RubyZip based IO wrapper with
ZipTricks, which enables support for non-seekable IO-like
objects and is needed for use cases like streaming the
XLSX file to a web client while it is still being generated.

Because the zip_tricks gem requires at least Ruby 2.1.0
due to the use of keyword arguments with defaule, we
also raise the xlsxtream required Ruby version to 2.1.0.